requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Strapped AI Dependencies
# ──────────────────────────────────────────────────────────────────────────────

# Microsoft Amplifier (core + foundation)
amplifier-core @ git+https://github.com/microsoft/amplifier-core@main
amplifier-foundation @ git+https://github.com/microsoft/amplifier-foundation@main

# Azure Functions runtime (for email processing function)
azure-functions>=1.21.0

# Azure Identity & Graph
azure-identity>=1.19.0
msgraph-sdk>=1.12.0

# Azure OpenAI
openai>=1.58.0

# PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data models
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Web app (landing page, auth, dashboard)
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=23.0.0
jinja2>=3.1.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dateutil>=2.9.0
pytz>=2024.2
python-dotenv>=1.0.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
